bimac {ouch}	R Documentation

Anolis bimaculatus lizard size data

Description

This is the Anolis bimaculatus dataset used in Butler & King (2004). It is used to test a hypothesis of character displacement using an interspecific dataset of body sizes and current data on sympatry/allopatry. The data frame has the following columns: species which are species names, size which is the phenotypic data, and the variables ancestor and time which specify the topology of the phylogeny and the location of the nodes in time, respectively. The columns OU.1, OU.3, OU.4, and OU.LP specify four hypothetical arrangements of selective regimes. Explanations of the data follow:

Body size.

We use the phenotypic data and phylogeny of Losos (1990), which employed the head lengths (of males) as a proxy for body size. In this group of lizards, head length correlates very strongly with snout-to-vent length and the cube root of mass, which are standard measures of body size. The data are head lengths in mm; note that we use the log of this value in analyses.

Tree topology

The tree topology is encoded via two vectors: ancestor and time. Each node of the' phylogenetic tree has a corresponding row in the data frame, numbered from 1 to 45. The columns ancestor and time specify the phylogeny. The ancestor variable specifies the topology: it is a list indicating the ancestor of each node. The root node has ancestor 0. The variable time specifies the temporal location of each node, with the root node being at time 0.

Specifications of selective regimes.

(Columns OU.1, OU.3, OU.4, OU.LP). These columns are factors, the levels of which correspond to the "paintings" of the respective adaptive regime hypotheses onto the phylogeny (see paint()). Each selective regime is named (small, medium, large, etc.). Each column corresponds to a different painting of the selective regimes, and thus to a different hypothesis. In this example, there are 3 alternative models (see Butler & King 2004): OU.4 is 4-regime model, OU.3 is 3-regime model (all ancestors are medium), OU.LP is the linear parsimony model.

Format

A data frame with 45 observations on the following 8 variables.

node

Labels for the nodes.

species

Species names for extant species.

size

Body size (head length in mm) of extant species.

ancestor

Ancestral node.

time

Time of node.

OU.1

a factor with levels ns

OU.3

a factor with levels small, medium, large

OU.4

a factor with levels small, medium, large, anc

OU.LP

a factor with levels small, medium, large

Author(s)

Marguerite A. Butler and Aaron A. King

Source

Butler, M.A. and A.A. King. 2004. Phylogenetic comparative analysis: a modeling approach for adaptive evolution. Am. Nat. 164:683–695.

References

Lazell, J. D. 1972. The anoles (Sauria: Iguanidae) of the Lesser Antilles. Bull. Mus. Comp. Zool., 143:1–115.

Losos, J. B. 1990. A phylogenetic analysis of character displacement in Caribbean Anolis lizards. Evolution, 44:558–569.

See Also

Other examples: anolis.ssd

Examples

## Analysis of Anolis bimaculatus data
tree <- with(bimac,ouchtree(node,ancestor,time/max(time),species))
plot(tree,node.names=TRUE)

h1 <- brown(log(bimac['size']),tree)
h1
plot(h1)

h2 <- hansen(log(bimac['size']),tree,bimac['OU.1'],sqrt.alpha=1,sigma=1)
h2
plot(h2)

h3 <- hansen(log(bimac['size']),tree,bimac['OU.3'],sqrt.alpha=1,sigma=1)
h3
plot(h3)

h4 <- hansen(log(bimac['size']),tree,bimac['OU.4'],sqrt.alpha=1,sigma=1)
h4
plot(h4)

h5 <- hansen(log(bimac['size']),tree,bimac['OU.LP'],sqrt.alpha=1,sigma=1,reltol=1e-5)
h5 <- update(h5,method='subplex',reltol=1e-11,parscale=c(0.1,0.1),hessian=TRUE)
h5

simdat <- simulate(h5,nsim=10)
hsim <- update(h5,data=simdat[[1]])
summary(hsim)
bsim <- update(h1,data=simdat[[1]])
summary(bsim)

[Package ouch version 2.16-5 Index]