summary {ouch}	R Documentation

Summary methods

Description

Summary methods

Usage

## S4 method for signature 'hansentree'
summary(object, ...)

## S4 method for signature 'browntree'
summary(object, ...)

Arguments

object
fitted model object.

...
additional arguments, ignored.

Value

summary applied to a hansentree method displays the estimated \alpha and \sigma^2 matrices as well as various quantities describing the goodness of model fit.

summary applied to a browntree object returns information about the fitted model, including parameter estimates and quantities describing the goodness of fit.

See Also

Other methods for ouch trees: as_data_frame, bootstrap(), coef(), logLik, ouch-package, paint(), plot(), print(), simulate(), update()

[Package ouch version 2.20-0 Index]