ape2ouch {ouch}	R Documentation

Convert an ape tree to an ouch tree

Description

ape2ouch translates ape's phylo representation of a phylogenetic tree into ouch's ouchtree representation. The user can change the branch lengths while preserving the topology.

Usage

ape2ouch(tree, scale = TRUE, branch.lengths = tree$edge.length)

Arguments

tree
a tree of class ape::phylo.

scale
optional. If scale=TRUE, the tree's depth will be scaled to 1. If scale is a number, then the branch lengths will be scaled by this number.

branch.lengths
optional vector of branch lengths.

Author(s)

A. A. King, D. Ackerly

See Also

ouchtree

[Package ouch version 2.16-7 Index]