bootstrap {ouch}	R Documentation

Bootstrapping for uncertainty quantification

Description

Generic bootstrapping for ouch models.

Usage

## S4 method for signature 'missing'
bootstrap(object, ...)

## S4 method for signature 'ANY'
bootstrap(object, ...)

## S4 method for signature 'browntree'
bootstrap(object, nboot = 200, seed = NULL, ...)

## S4 method for signature 'hansentree'
bootstrap(object, nboot = 200, seed = NULL, ...)

Arguments

object
A fitted model object.

...
Additional arguments are passed to update.

nboot
integer; number of bootstrap replicates.

seed
integer; setting seed to a non-NULL value allows one to fix the random seed (see simulate).

Details

bootstrap performs a parametric bootstrap for estimation of confidence intervals.

See Also

Other methods: coef(), logLik, plot(), print(), simulate(), summary(), update()

[Package ouch version 2.16-2 Index]