coef,browntree-method {ouch}	R Documentation

Model coefficients

Description

coef extracts the parameters from a fitted model object.

Usage

## S4 method for signature 'browntree'
coef(object, ...)

## S4 method for signature 'hansentree'
coef(object, ...)

Arguments

object
fitted model object.

...
additional arguments, ignored.

Value

coef applied to a browntree object extracts a list with three elements:

sigma

the coefficients of the sigma matrix.

theta

a list of the estimated optima, one per character.

sigma..sq.matrix

the sigma-squared matrix itself.

coef applied to a hansentree object returns a named list containing the estimated alpha and sigma^2 matrices(given as the alpha.matrix and sigma.sq.matrix elements, respectively) but also the MLE returned by the optimizer (as sqrt.alpha and sigma, respectively). The latter elements should not be interpreted, but can be used to restart the algorithm, etc.

[Package ouch version 2.15-1 Index]