ouchtree {ouch}	R Documentation

Phylogenetic tree object in 'ouch' format.

Description

An object containing a phylogenetic tree in a form suitable for using ouch methods.

Usage

ouchtree(nodes, ancestors, times, labels = as.character(nodes))

Arguments

nodes
A character vector giving the name of each node. These are used internally and must be unique.

ancestors
Specification of the topology of the phylogenetic tree. This is in the form of a character vector specifying the name (as given in the nodes argument) of the immediate ancestor of each node. In particular, the i-th name is that of the ancestor of the i-th node. The root node is distinguished by having no ancestor (i.e., NA).

times
A vector of nonnegative numbers, one per node in the tree, specifying the time at which each node is located. Time should be increasing from the root node to the terminal twigs.

labels
Optional vector of node labels. These will be used in plots to label nodes. It is not necessary that these be unique.

Details

ouchtree creates an ouchtree object. This contains the topology, branch times, and epochs. It also (optionally) holds names of taxa for display purposes.

Author(s)

Aaron A. King

See Also

ape2ouch

Other phylogenetic comparative models: brown(), hansen()

Examples

tree <- with(
  bimac,
  ouchtree(nodes=node,ancestors=ancestor,times=time,labels=species)
)
tree

plot(tree)
plot(tree,node.names=TRUE)

[Package ouch version 2.16-3 Index]