ouch-package {ouch}	R Documentation

Ornstein-Uhlenbeck methods for comparative phylogenetic hypotheses

Description

The ouch package provides facilities for phylogenetic comparative analysis based on Ornstein-Uhlenbeck models of trait evolution along a phylogeny. Multivariate data and complex adaptive hypotheses are supported.

Classes

The basic class, ouchtree, is provided to encode a phylogenetic tree. Plot and print methods are provided.

The class browntree derives from class ouchtree and encodes the results of fitting a Brownian Motion model to data.

The class hansentree derives from class ouchtree and encodes the results of fitting a Hansen model to data.

Detailed Documentation

Phylogenies in ouch format: ouchtree(), ape2ouch()

Brownian motion models: brown()

Ornstein-Uhlenbeck models: hansen(), paint()

Simulation of models: simulate()

Display of data: plot()

Extraction of information from fitted models: summary(), logLik(), coef()

Example datasets: anolis.ssd, bimac

Author(s)

Aaron A. King

References

Butler, M.A. and A.A. King. 2004. Phylogenetic comparative analysis: a modeling approach for adaptive evolution. Am. Nat. 164:683–695.

Cressler, C. E., Butler, M. A., and King, A. A. 2015. Detecting adaptive evolution in phylogenetic comparative analysis using the Ornstein-Uhlenbeck model. Syst. Biol., 64:953–968.

See Also

Other phylogenetic comparative models: brown(), hansen(), ouchtree, paint()

Other methods for ouch trees: bootstrap(), coef(), logLik, paint(), plot(), print(), simulate(), summary(), update()

Other examples: anolis.ssd, bimac

[Package ouch version 2.16-9 Index]