plot {ouch}	R Documentation

ouch plotting functions

Description

Plot phylogenetic trees, with or without regime paintings.

Usage

## S4 method for signature 'ouchtree'
plot(
  x,
  y,
  ...,
  regimes = NULL,
  ladderize = TRUE,
  node.names = FALSE,
  legend = TRUE,
  labels,
  frame.plot = FALSE,
  palette = rainbow,
  margin = 0.1,
  text_opts = list(),
  legend_opts = list()
)

## S4 method for signature 'hansentree'
plot(x, ..., regimes)

Arguments

x
object to plot.

y
ignored.

...
additional arguments, passed to plot.

regimes
factor or character; a vector of regime paintings.

ladderize
logical; should the tree be ladderized?

node.names
logical; should node names be displayed?

legend
logical; display a legend?

labels
character; taxon labels.

frame.plot
a logical indicating whether a box should be drawn around the plot.

palette
function or character; specifies the colors to be used for the several regimes on the tree. Specified as a function, when given an integer, n, the function should create a vector of n colors. See, for example rainbow. One can also specify the n colors as a vector of color codes. There must be at least as many colors as levels in the regimes.

margin
numeric; width of the right margin (as a fraction of the plot width). Adjust this if labels are clipped. If different left and right margins are desired, furnish two numbers here.

text_opts
options for the labels; passed to text

legend_opts
options for the the legend; passed to legend

See Also

Other methods: bootstrap(), coef(), logLik, print(), simulate(), summary(), update()

[Package ouch version 2.16-1 Index]