print {ouch}	R Documentation

Print and show methods

Description

Print and show methods

Usage

## S4 method for signature 'ouchtree'
print(x, ...)

## S4 method for signature 'ouchtree'
show(object)

## S4 method for signature 'browntree'
show(object)

## S4 method for signature 'browntree'
print(x, ...)

## S4 method for signature 'hansentree'
print(x, ...)

## S4 method for signature 'hansentree'
show(object)

Arguments

...
additional arguments, ignored.

object, x
object to display.

Value

print returns x, invisibly.

show returns NULL, invisibly.

print displays the tree as a table, along with the coefficients of the fitted model and diagnostic information.

print displays the tree as a table, along with the coefficients of the fitted model and diagnostic information.

See Also

Other methods: bootstrap(), coef(), logLik, plot(), simulate(), summary(), update()

[Package ouch version 2.16-4 Index]