summary {ouch}	R Documentation

ouch summary methods

Description

ouch summary methods

Usage

## S4 method for signature 'browntree'
summary(object, ...)

## S4 method for signature 'hansentree'
summary(object, ...)

Arguments

object
fitted model object.

...
additional arguments, ignored.

Value

summary applied to a browntree object returns information about the fitted model, including parameter estimates and quantities describing the goodness of fit.

summary applied to a hansentree method displays the estimated alpha and sigma^2 matrices as well as various quantities describing the goodness of model fit.

See Also

Other methods for ouch trees: bootstrap(), coef(), logLik, ouch-package, paint(), plot(), print(), simulate(), update()

[Package ouch version 2.16-9 Index]